ultimata la velocità sui bordi e bilanciate le velocità

+++ simulation.hpp
@@ -8,7 +8,7 @@
 
 void update(Flock& flock, int steps_per_evolution, sf::Time delta_t);
 
-void evolve(Flock& stormo, double delta_t);
+void evolve(Flock& stormo, float delta_t);
 
 Boid solve(Flock& stormo, double delta_t, Boid& boid);
